+++ backend/crates/entity/src/lib.rs
@@ -2,4 +2,5 @@
 
 pub mod organization;
 pub mod permission;
+pub mod sector;
 pub mod user;